save delete kategori mading

+++ app/Http/Controllers/Admin/KelolaKategoriController.php
@@ -5,17 +5,12 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\KelolaKategoriRequest;
 use App\KelolaKategori;
+use App\KelolaMading;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
 class KelolaKategoriController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-
     //  fungsi memunculkan data dari halaman utama kelola kategori
     public function index(Request $request)
     {
@@ -32,24 +27,11 @@ class KelolaKategoriController extends Controller
         ]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-
     //  fungsi membuat data di kelola kategori
     public function create()
     {
         return view('pages.admin.kelola-kategori.create');
     }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
 
     //  fungsi yang mengarahkan untuk melakukan penyimpanan data
     public function store(KelolaKategoriRequest $request)
@@ -61,23 +43,10 @@ class KelolaKategoriController extends Controller
         return redirect()->route('kelola-kategori.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         //
     }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
 
     //  fungsi yang mengarahkan ke halaman edit
     public function edit($id)
@@ -88,14 +57,6 @@ class KelolaKategoriController extends Controller
             'item' => $item
         ]);
     }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
 
     //  fungsi untuk menyimpan data terbaru
     public function update(KelolaKategoriRequest $request, $id)
@@ -109,19 +70,18 @@ class KelolaKategoriController extends Controller
         return redirect()->route('kelola-kategori.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-
     // Fungsi untuk menghapus data
     public function destroy($id)
     {
-        $item = KelolaKategori::findOrFail($id);
-        $item->delete();
-
-        return redirect()->route('kelola-kategori.index');
+        // cek apakah kategori sudah dipakai di mading
+        if (KelolaMading::where('kelola_kategori_id', '=', $id)->exists()) {
+            // tampilkan pesan error
+            return response()->json(['success'=> false]);
+        }else{
+            // save delete, karena kategori tidak dipakai
+            $item = KelolaKategori::findOrFail($id);
+            $item->delete();
+            return response()->json(['success'=> true]);
+        }
     }
 }
